code cleanup after prs #2433 #2434 #2428
