+++ karate-core/src/test/java/com/intuit/karate/core/jscall2/all.feature
@@ -1,16 +1,15 @@
 Feature: responseStatus variable tests
-  Scenario: config js test
-    Given url 'https://www.baidu.com'
-    When method GET
-    Then status 200
 
+  Scenario: config js test
+    * url serverUrl
+    * method get
+    * status 200
     * print 'responseStatus: ' + responseStatus
     * assert isResponseStatus200_config()
 
   Scenario: call test
-    Given url 'https://www.baidu.com'
-    When method GET
-    Then status 200
-
+    * url serverUrl
+    * method get
+    * status 200
     * print 'responseStatus: ' + responseStatus
     * assert isResponseStatus200_call()
