+++ karate-core/src/test/java/com/intuit/karate/core/jscall2/call-single.feature
@@ -1,5 +1,3 @@
-
-#@ignore
 Feature:: common setup
 
   Scenario: common setup
