+++ karate-core/src/test/java/com/intuit/karate/core/jscall2/mock.feature
@@ -1,0 +1,4 @@
+Feature:
+
+Scenario:
+* def response = { success: true }
